+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Unreleased
 
 * Demo: Prevent spaces in theme-info link
+* Fix newtheme-popos.sh script location detection
 
 ### 9.4.5: 2024-10-25